pyproject.toml
[tool.poetry]
name = "remissbanken-django"
version = "0.1.0"
description = ""
authors = ["Hassan <hassanakily@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.13.1"
djongo = "^1.3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
